back-end/moni/requirements.txt
Django==3.1.7
dj-database-url==0.5.0
django-import-export
django-cors-headers==3.5.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
whitenoise==5.2.0
requests==2.24.0